backend/requirements.txt
flask
flask-cors
python-dotenv
openai
requests
beautifulsoup4 